# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Currency Exchange API",
        "version": "1.0.0",
        "description": "API for managing currencies and exchange rates"
    },
    "servers": [
        {
            "url": "http://localhost:3000/api/v1"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "summary": "Login",
                "description": "Authenticate user and return access token",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated successfully",
                        "headers": {
                            "Authorization": {
                                "schema": {
                                    "type": "string"
                                },
                                "description": "Access token"
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid email or password"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "summary": "Logout",
                "description": "Logout the user",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                }
            }
        },
        "/currencies": {
            "get": {
                "summary": "Get all currencies",
                "description": "Fetch a list of all available currencies",
                "responses": {
                    "200": {
                        "description": "List of currencies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Currency"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add a currency",
                "description": "Add a new currency",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Currency added successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/currencies/{id}/history": {
            "get": {
                "summary": "Get currency history",
                "description": "Fetch the history of a currency's exchange rates",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "fromDate",
                        "in": "query",
                        "description": "Start date for exchange rate history",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "toDate",
                        "in": "query",
                        "description": "End date for exchange rate history",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency history",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExchangeRate"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/currencies/edit/{id}": {
            "put": {
                "summary": "Edit a currency",
                "description": "Update the details of a currency",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Currency updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/currencies/delete/{id}": {
            "delete": {
                "summary": "Delete a currency",
                "description": "Delete a currency by ID",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/exchange-rates": {
            "get": {
                "summary": "Get exchange rates for today",
                "description": "Fetch today's exchange rates",
                "responses": {
                    "200": {
                        "description": "List of exchange rates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExchangeRate"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add a new exchange rate",
                "description": "Add a new exchange rate",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "fromCurrency": {
                                        "type": "integer"
                                    },
                                    "toCurrency": {
                                        "type": "integer"
                                    },
                                    "rate": {
                                        "type": "number"
                                    }
                                },
                                "required": [
                                    "date",
                                    "fromCurrency",
                                    "toCurrency",
                                    "rate"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exchange rate added successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/exchange-rate/edit/{id}": {
            "put": {
                "summary": "Edit an exchange rate",
                "description": "Update the details of an exchange rate",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Exchange rate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "rate": {
                                        "type": "number"
                                    }
                                },
                                "required": [
                                    "rate"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exchange rate updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/exchange-rates/delete/{id}": {
            "delete": {
                "summary": "Delete an exchange rate",
                "description": "Delete an exchange rate by ID",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Exchange rate ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange rate deleted successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Currency": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
            "ExchangeRate": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "fromCurrency": {
                        "type": "integer"
                    },
                    "toCurrency": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string",
                        "format": "date"
                    },
                    "rate": {
                        "type": "number"
                    }
                },
                "required": [
                    "id",
                    "fromCurrency",
                    "toCurrency",
                    "date",
                    "rate"
                ]
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}